add user to book model who create the book object

+++ book/models.py
@@ -4,11 +4,10 @@ from django.contrib.auth import get_user_model
 
 
 class Book(models.Model):
-    #cmmm
     title = models.CharField(max_length=200)
     author = models.CharField(max_length=200)
     description = models.TextField()
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.PositiveIntegerField()
     cover = models.ImageField(upload_to='covers/', blank=True)
     user = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
